- Funções de API para buscar, substituir e para apagar pontos individuais - Definição de uma rota separada para obtenção de pontos

+++ DataFormats.cs
@@ -3,8 +3,8 @@
     public class Point3D
     {
         public int ID { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
     }
 }
